+++ test/cpp/catch2/catch.hpp
@@ -1,5 +1,0 @@
-#ifndef CATCH_CONFIG_MAIN
-#define CATCH_CONFIG_MAIN
-#endif
-
-#include "catch.hpp.txt"